% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ann_expr.R
\name{add_row}
\alias{add_row}
\title{add_row}
\usage{
add_row(data, add_row, rowname)
}
\arguments{
\item{data}{data}

\item{add_row}{a new row}

\item{rowname}{rowname}
}
\value{
a data.frame
}
\description{
Add a row
}
\author{
Yuanlong Hu
}
