% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ann_expr.R
\name{ann_expr}
\alias{ann_expr}
\title{ann_expr}
\usage{
ann_expr(data, ann_db, ann_df, Stat = c("max", "mean", "median", "min", "IQR"))
}
\arguments{
\item{data}{data}

\item{ann_db}{The name of the annotation package}

\item{ann_df}{gene symbol data frame}

\item{Stat}{"max","mean","median","min","IQR"}
}
\value{
a data.frame
}
\description{
Convert probe ID to gene symbol
}
\author{
Yuanlong Hu
}
