% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ann_expr.R
\name{ann_expr2}
\alias{ann_expr2}
\title{ann_expr2}
\usage{
ann_expr2(
  data,
  GPL,
  probe_symbol = c("NAME", "GENE_SYMBOL"),
  Stat = c("max", "mean", "median", "min", "IQR")
)
}
\arguments{
\item{data}{data}

\item{GPL}{GPL file}

\item{probe_symbol}{probe id and gene symbol}

\item{Stat}{"max","mean","median","min","IQR"}
}
\value{
a data.frame
}
\description{
Convert probe ID to gene symbol using GPL
}
\author{
Yuanlong Hu
}
