% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_apr.R
\name{AssoAnalysis}
\alias{AssoAnalysis}
\title{Run the TCMshiny Application}
\usage{
AssoAnalysis()
}
\arguments{
\item{...}{A series of options to be used inside the app.}
}
\description{
Run the TCMshiny Application
}
\author{
Yuanlong Hu
}
