% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.R
\name{enrich_geneset}
\alias{enrich_geneset}
\title{enrich_geneset}
\usage{
enrich_geneset(
  genes,
  pvalueCutoff = 0.05,
  qvalueCutoff = 0.05,
  kegg_internal_data = FALSE,
  GMTset = NULL,
  useGMTset = FALSE,
  IDtoNAME = NA
)
}
\arguments{
\item{genes}{res}

\item{pvalueCutoff}{p-value cutoff}

\item{qvalueCutoff}{q-value Cutoff}

\item{kegg_internal_data}{Use kegg internal data}

\item{GMTset}{gmt file}

\item{useGMTset}{use geneset from gmt file}

\item{IDtoNAME}{ID to name}
}
\value{
a gseaResult object
}
\description{
Over-representation test
}
\author{
Yuanlong Hu
}
