% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.R
\name{enrich_gsea}
\alias{enrich_gsea}
\title{enrich_gsea}
\usage{
enrich_gsea(res, pvalueCutoff = 0.05, kegg_internal_data = FALSE)
}
\arguments{
\item{res}{res}

\item{pvalueCutoff}{p-value cutoff}

\item{kegg_internal_data}{Use kegg internal data}
}
\value{
a ggplot2 object
}
\description{
gsea
}
\author{
Yuanlong Hu
}
