% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findDEG.R
\name{findDEG}
\alias{findDEG}
\title{findDEG}
\usage{
findDEG(data, pdata, contrasts)
}
\arguments{
\item{data}{A matrix of expression values where rows correspond to genes and columns correspond to samples.}

\item{pdata}{A character vector of phenotype.}

\item{contrasts}{character vector specifying contrasts}
}
\value{
A list
}
\description{
Calculate differentially expressed genes using limma method.
}
\author{
Yuanlong Hu
}
