% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analysis.R
\name{graph_node}
\alias{graph_node}
\title{graph_node}
\usage{
graph_node(data, input = c("df", "adjmat"), weighted = FALSE)
}
\arguments{
\item{data}{data}

\item{input}{"df","adjmat"}

\item{weighted}{weighted}
}
\value{
a data frame
}
\description{
graph node analysis
}
\author{
Yuanlong Hu
}
