% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_expression.R
\name{MIC_pvalue}
\alias{MIC_pvalue}
\title{MIC_pvalue}
\usage{
MIC_pvalue(x, MIC_mat, n)
}
\arguments{
\item{x}{data matrix}

\item{MIC_mat}{MIC matrix}

\item{n}{n}
}
\value{
p-value matrix
}
\description{
plot cor
}
\author{
Yuanlong Hu
}
