% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway.R
\name{pathway_gsva}
\alias{pathway_gsva}
\title{pathway_gsva}
\usage{
pathway_gsva(
  data,
  geneset,
  gmt = NULL,
  method = c("ssgsea", "gsva", "zscore", "plage"),
  rna_seq = FALSE
)
}
\arguments{
\item{data}{a matrix of expression values where rows correspond to genes and columns correspond to samples.}

\item{geneset}{a list or dataframe}

\item{gmt}{gmtfile}

\item{method}{one of "ssgsea","gsva", "zscore",and "plage"}

\item{rna_seq}{rna-seq data or microarray data}
}
\value{
a data.frame
}
\description{
Estimates GSVA enrichment scores.
}
\author{
Yuanlong Hu
}
