% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdata.R
\name{pdataSummary}
\alias{pdataSummary}
\title{pdataSummary}
\usage{
pdataSummary(data, group, show_p = T)
}
\arguments{
\item{data}{pdata}

\item{group}{group}

\item{show_p}{show p}
}
\value{
a summary data frame
}
\description{
pdataSummary
}
\author{
Yuanlong Hu
}
