% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_cor}
\alias{plot_cor}
\title{plot_cor}
\usage{
plot_cor(
  expr,
  gene.x,
  gene.y,
  geom_smooth = TRUE,
  method = "lm",
  group = NULL,
  group_point = TRUE,
  group_smooth = TRUE
)
}
\arguments{
\item{expr}{a matrix of expression values where rows correspond to genes and columns correspond to samples.}

\item{gene.x}{x}

\item{gene.y}{y}

\item{geom_smooth}{rna-seq data or microarray data}

\item{method}{one of "lm", "loess", "glm"}

\item{group}{group}

\item{group_point}{group_point}

\item{group_smooth}{group_smooth}
}
\value{
a ggplot
}
\description{
plot point
}
\author{
Yuanlong Hu
}
