% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_expression.R
\name{plot_point}
\alias{plot_point}
\title{plot_point}
\usage{
plot_point(
  data,
  x,
  y,
  group = NULL,
  point_group = TRUE,
  geom_smooth = TRUE,
  method = c("lm", "glm", "gam", "loess"),
  smooth_group = TRUE
)
}
\arguments{
\item{data}{expr data}

\item{x}{x}

\item{y}{y}

\item{group}{group}

\item{point_group}{logical}

\item{geom_smooth}{logical}

\item{method}{c("lm","glm","gam","loess")}

\item{smooth_group}{logical}
}
\value{
ggplot object
}
\description{
plot cor
}
\author{
Yuanlong Hu
}
