% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_fit.R
\name{plotBorutaImpHistory}
\alias{plotBorutaImpHistory}
\title{plotBorutaImpHistory}
\usage{
plotBorutaImpHistory(
  res,
  select = NULL,
  select2 = c("Confirmed", "Rejected", "Tentative")
)
}
\arguments{
\item{res}{the result of Boruta}

\item{select}{a vector.}
}
\value{
a Boruta object
}
\description{
plot Boruta ImpHistory
}
\author{
Yuanlong Hu
}
