% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_expression.R
\name{plotCorHeatmap}
\alias{plotCorHeatmap}
\title{plotCorHeatmap}
\usage{
plotCorHeatmap(res, x, y, pvalueCutoff = 0.05)
}
\arguments{
\item{res}{Hmis}

\item{x}{x}

\item{y}{y}

\item{pvalueCutoff}{p-value cutoff}
}
\value{
ggplot object
}
\description{
plot cor
}
\author{
Yuanlong Hu
}
