% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_fit.R
\name{plotExprBox}
\alias{plotExprBox}
\title{plotExprBox}
\usage{
plotExprBox(expr, select, pdata, comparisons = list(c("C1", "C2")))
}
\arguments{
\item{expr}{the expr data}

\item{select}{a vector.}

\item{pdata}{pdata}

\item{comparisons}{a list}
}
\value{
a Boruta object
}
\description{
plot ExprBox
}
\author{
Yuanlong Hu
}
