% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_fit.R
\name{plotExprBox2}
\alias{plotExprBox2}
\title{plotExprBox2}
\usage{
plotExprBox2(
  expr,
  select,
  pdata,
  comparisons = list(c("S1", "S2")),
  label = c("p.signif", "p.format"),
  ncol = 2,
  nrow = 2
)
}
\arguments{
\item{expr}{the expr data}

\item{select}{a vector.}

\item{pdata}{pdata}

\item{comparisons}{a list}

\item{label}{"p.signif","p.format"}

\item{ncol}{the number of col}

\item{nrow}{the number of now}
}
\value{
a ggplot2 object
}
\description{
plot ExprBox2
}
\author{
Yuanlong Hu
}
