% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_fit.R
\name{plotExprDIM}
\alias{plotExprDIM}
\title{plotExprDIM}
\usage{
plotExprDIM(
  expr,
  feature,
  pdata = NULL,
  method,
  addEllipses = TRUE,
  ellipse_type = "confidence",
  ellipse_level = 0.95,
  tsne_perplexity = 30,
  ...
)
}
\arguments{
\item{expr}{expr}

\item{feature}{feature}

\item{pdata}{a vector.}

\item{method}{"PCA" or "tSNE"}

\item{addEllipses}{TRUE or FALSE}

\item{ellipse_type}{"convex" or "confidence"}

\item{ellipse_level}{0.95}

\item{tsne_perplexity}{perplexity}

\item{...}{other}
}
\value{
a ggplot2 object
}
\description{
plot PCA and t-sne
}
\author{
Yuanlong Hu
}
