% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeneset.R
\name{plotExprGSEAHeatmap}
\alias{plotExprGSEAHeatmap}
\title{plotExprGSEAHeatmap}
\usage{
plotExprGSEAHeatmap(
  res_gsea,
  res_deg,
  select,
  sort = "logFC",
  logFCCutoff = 0.1,
  pvalueCutoff = 0.05,
  adjpvalueCutoff = 0.1,
  show_NES = TRUE
)
}
\arguments{
\item{res_gsea}{GSEA result}

\item{res_deg}{DEGs result}

\item{select}{selected ID}

\item{sort}{sort method}

\item{logFCCutoff}{abs logFC cutoff}

\item{pvalueCutoff}{p-value cutoff}

\item{adjpvalueCutoff}{adjusted p-value cutoff}

\item{show_NES}{Whether to show GSEA NES}
}
\value{
a ggplot2 object
}
\description{
Heatmap displaying gene expression in GSEA
}
\author{
Yuanlong Hu
}
