% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_fit.R
\name{plotExprPCA}
\alias{plotExprPCA}
\title{plotExprPCA}
\usage{
plotExprPCA(expr, feature, pdata)
}
\arguments{
\item{expr}{expr}

\item{feature}{feature}

\item{pdata}{a vector.}
}
\value{
a ggplot2 object
}
\description{
plot PCA
}
\author{
Yuanlong Hu
}
