% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_fit.R
\name{plotExprVolcano}
\alias{plotExprVolcano}
\title{plotExprVolcano}
\usage{
plotExprVolcano(
  res,
  selectlabels = NULL,
  logFCcutoff = 1,
  xlab = "log FoldChange",
  ylab = "-log10 adjusted P-value"
)
}
\arguments{
\item{res}{the result of limma}

\item{selectlabels}{a vector.}

\item{logFCcutoff}{logFC cutoff}

\item{xlab}{a charact}

\item{ylab}{a charact}
}
\value{
a Boruta object
}
\description{
plot ExprVolcano
}
\author{
Yuanlong Hu
}
