% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_fit.R
\name{plotForest}
\alias{plotForest}
\title{plotForest}
\usage{
plotForest(
  res,
  select = NULL,
  cutoff = 1,
  pCutoff = 0.05,
  point = "HR",
  xmin = "CI_lower",
  xmax = "CI_upper",
  pvalue = "p.value"
)
}
\arguments{
\item{res}{result}

\item{select}{selected object}

\item{cutoff}{cutoff}

\item{pCutoff}{p-value Cutoff}

\item{point}{column point}

\item{xmin}{column xmin}

\item{xmax}{column xmax}

\item{pvalue}{column pvalue}
}
\value{
a ggplot2 object
}
\description{
plot Forest
}
\author{
Yuanlong Hu
}
