% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_fit.R
\name{plotGroupBar}
\alias{plotGroupBar}
\title{plotGroupBar}
\usage{
plotGroupBar(pdata, x, fill)
}
\arguments{
\item{pdata}{pdata}

\item{x}{a vector.}

\item{fill}{logFC cutoff}
}
\value{
a ggplot2 object
}
\description{
plot GroupBar
}
\author{
Yuanlong Hu
}
