% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findDEG.R
\name{plotSummaryDEGHeatmap}
\alias{plotSummaryDEGHeatmap}
\title{plotSummaryDEGHeatmap}
\usage{
plotSummaryDEGHeatmap(
  reslist,
  select,
  pvalueCutoff = 1,
  adjustpCutoff = 1,
  marker = F
)
}
\arguments{
\item{reslist}{the list of results}

\item{select}{selected genes}

\item{pvalueCutoff}{pvalue Cutoff}

\item{adjustpCutoff}{adjust p Cutoff}

\item{marker}{P marker}
}
\value{
ggplot2
}
\description{
Heatmap showing the summary of DEGs
}
\author{
Yuanlong Hu
}
