% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeneset.R
\name{plotSummaryGSEAHeatmap}
\alias{plotSummaryGSEAHeatmap}
\title{plotSummaryGSEAHeatmap}
\usage{
plotSummaryGSEAHeatmap(reslist, selectID = NULL, show_top = 20)
}
\arguments{
\item{reslist}{GSEA result list}

\item{selectID}{selected ID}

\item{show_top}{top}
}
\value{
a ggplot2 object
}
\description{
Summary NES in GSEA results
}
\author{
Yuanlong Hu
}
