% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MR.R
\name{runTwoSampleMR2}
\alias{runTwoSampleMR2}
\title{runTwoSampleMR2}
\usage{
runTwoSampleMR2(
  expID,
  outID,
  p1 = 5e-08,
  r2 = 0.01,
  kb = 1000,
  maf_threshold = 0.01,
  method_list = subset(mr_method_list(), use_by_default)$obj,
  proxies = TRUE
)
}
\arguments{
\item{expID}{IDs}

\item{outID}{Outcome IDs}

\item{p1}{Significance threshold.}

\item{r2}{Clumping r2 cut off.}

\item{kb}{Clumping distance cutoff.}

\item{maf_threshold}{MAF threshold to try to infer palindromic SNPs.}

\item{method_list}{List of methods to use in analysis.}

\item{proxies}{Look for LD tags?}
}
\value{
a data frame object
}
\description{
run TwoSampleMR
}
\author{
Yuanlong Hu
}
