% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_fit.R
\name{select_lm_single}
\alias{select_lm_single}
\title{select_lm_single}
\usage{
select_lm_single(expr, pdata)
}
\arguments{
\item{expr}{a matrix of expression values where rows correspond to genes and columns correspond to samples.}

\item{pdata}{a vector.}
}
\value{
a data.frame
}
\description{
Fit single lm methods
}
\author{
Yuanlong Hu
}
