% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_fit.R
\name{select_logistic_single}
\alias{select_logistic_single}
\title{select_logistic_single}
\usage{
select_logistic_single(expr, feature = NULL, pdata, adjust_p_method = "BH")
}
\arguments{
\item{expr}{a matrix of expression values where rows correspond to genes and columns correspond to samples.}

\item{feature}{feature}

\item{pdata}{a vector.}

\item{adjust_p_method}{adjust p-value}
}
\value{
a data.frame
}
\description{
Fit single logistic methods
}
\author{
Yuanlong Hu
}
