% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_expression.R
\name{sim_expr}
\alias{sim_expr}
\title{sim_expr}
\usage{
sim_expr(
  data,
  gene,
  method = c("pearson", "spearman", "MIC"),
  MIC_pvalue = FALSE,
  pcutoff = 0.05,
  cut = 2,
  n = 100
)
}
\arguments{
\item{data}{data}

\item{gene}{gene}

\item{method}{one of pearson and spearman.}

\item{MIC_pvalue}{MIC p-value}

\item{pcutoff}{<0.05}

\item{cut}{2}

\item{n}{10}
}
\value{
a data.frame
}
\description{
Calculate the similarity
}
\author{
Yuanlong Hu
}
