% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_expression.R
\name{sim_expr2}
\alias{sim_expr2}
\title{sim_expr}
\usage{
sim_expr2(..., expr = NULL, pdata = NULL, method = c("pearson", "spearman"))
}
\arguments{
\item{method}{one of pearson and spearman.}

\item{data}{data}

\item{gene}{gene}

\item{MIC_pvalue}{MIC p-value}

\item{pcutoff}{<0.05}

\item{cut}{2}

\item{n}{10}
}
\value{
a data.frame
}
\description{
Calculate the similarity
}
\author{
Yuanlong Hu
}
