% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.R
\name{enrich_gsea}
\alias{enrich_gsea}
\title{enrich_gsea}
\usage{
enrich_gsea(
  res,
  pvalueCutoff = 0.05,
  kegg_internal_data = FALSE,
  GMTset = NULL,
  useGMTset = FALSE,
  IDtoNAME = NA,
  na.omit = TRUE
)
}
\arguments{
\item{res}{res}

\item{pvalueCutoff}{p-value cutoff}

\item{kegg_internal_data}{Use kegg internal data}

\item{GMTset}{gmt file}

\item{useGMTset}{use geneset from gmt file}

\item{IDtoNAME}{ID to name}

\item{na.omit}{Remove NA}
}
\value{
a gseaResult object
}
\description{
gsea
}
\author{
Yuanlong Hu
}
