% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeneset.R
\name{plotExprGenesetHeatmap}
\alias{plotExprGenesetHeatmap}
\title{plotExprGenesetHeatmap}
\usage{
plotExprGenesetHeatmap(
  res,
  select,
  selectgene = NULL,
  genesetlist,
  logFCCutoff = 0.2,
  pvalueCutoff = 0.05,
  adjpvalueCutoff = 0.05
)
}
\arguments{
\item{res}{result}

\item{select}{selected geneset name}

\item{genesetlist}{a geneset list}

\item{logFCCutoff}{abs logFC cutoff}

\item{pvalueCutoff}{p-value cutoff}

\item{adjpvalueCutoff}{adjusted p-value cutoff}
}
\value{
a ggplot2 object
}
\description{
plot geneset heatmap
}
\author{
Yuanlong Hu
}
