% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_fit.R
\name{plotExprPCA}
\alias{plotExprPCA}
\title{plotExprPCA}
\usage{
plotExprPCA(
  expr,
  feature,
  pdata = NULL,
  addEllipses = TRUE,
  ellipse_type = "confidence",
  ellipse_level = 0.95,
  ...
)
}
\arguments{
\item{expr}{expr}

\item{feature}{feature}

\item{pdata}{a vector.}

\item{addEllipses}{TRUE or FALSE}

\item{ellipse_type}{"convex" or "confidence"}

\item{ellipse_level}{0.95}

\item{...}{other}
}
\value{
a ggplot2 object
}
\description{
plot PCA
}
\author{
Yuanlong Hu
}
