% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_fit.R
\name{plotExprPCA}
\alias{plotExprPCA}
\title{plotExprPCA}
\usage{
plotExprPCA(expr, feature, pdata, addEllipses = TRUE, ellipse.type = "")
}
\arguments{
\item{expr}{expr}

\item{feature}{feature}

\item{pdata}{a vector.}

\item{addEllipses}{TRUE or FALSE}

\item{ellipse.type}{"convex"，"confidence"}
}
\value{
a ggplot2 object
}
\description{
plot PCA
}
\author{
Yuanlong Hu
}
