% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_fit.R
\name{select_Boruta}
\alias{select_Boruta}
\title{select_Boruta}
\usage{
select_Boruta(expr, pdata, pValue = 0.01, maxRuns = 100)
}
\arguments{
\item{expr}{a matrix of expression values where rows correspond to genes and columns correspond to samples.}

\item{pdata}{a vector.}

\item{pValue}{confidence level}

\item{maxRuns}{maximal number of importance source runs.}
}
\value{
a Boruta object
}
\description{
Select by Boruta
}
\author{
Yuanlong Hu
}
