% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_fit.R
\name{select_cox_single}
\alias{select_cox_single}
\title{select_cox_single}
\usage{
select_cox_single(expr, time, status, digits = 2)
}
\arguments{
\item{expr}{a matrix of expression values where rows correspond to genes and columns correspond to samples.}

\item{time}{a vector.}

\item{status}{status}

\item{digits}{digits}
}
\value{
a data.frame
}
\description{
Fit single cox methods
}
\author{
Yuanlong Hu
}
